pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "falcon-policy-scoring"
version = "1.5.0"
description = "CrowdStrike Falcon policy analysis and grading tool"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "CrowdStrike" }]
dependencies = [
    "crowdstrike-falconpy>=1.5.4",
    "tinydb>=4.8.0",
    "schedule>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["packaging", "rich"]
test = ["pytest>=8.3.5", "pytest-cov", "vcrpy>=7.0", "jsonschema"]
docs = ["sphinx>=8.2.3", "sphinx-autodoc-typehints>=3.2.0"]

[project.scripts]
policy-audit = "falcon_policy_scoring.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
